Break out tests for standalone methods in ingester module
